Initial refactor of summary page
